Extract backend-neutral cosine genotype scorer
